feat(worker-agent): bind channels to runtime threads
