+++ apps/worker-agent/src/tui.ts
@@ -1,116 +1,37 @@
 import { mkdir } from "node:fs/promises";
-import { homedir } from "node:os";
 import { join, resolve } from "node:path";
-import { argv, env, stdin, stdout } from "node:process";
+import { argv, stdin, stdout } from "node:process";
 import { createInterface } from "node:readline/promises";
 import { pathToFileURL } from "node:url";
 import type { AgentHost } from "@minpeter/pss-runtime";
 import { createNodeFileThreadHost } from "@minpeter/pss-runtime/node";
-import { z } from "zod";
-
-import { createConfiguredAgent, type WorkerAgentModelEnv } from "./agent";
+import { createConfiguredAgent } from "./agent";
 import type { WorkerAgentDeliveryResponse } from "./agent-do-delivery";
 import { threadStoreForHost } from "./agent-host-thread-store";
-import { type ChannelAddress, channelKey } from "./channel";
+import { localChannelBinding } from "./channel";
 import {
   createSessionIndexStore,
   type SessionIndexStore,
 } from "./session-index";
 import { createFileSessionIndexRepository } from "./session-index-node";
 import { createThreadStoreSessionTranscriptReader } from "./session-transcript";
+import {
+  resolveWorkerAgentTuiConfig,
+  type WorkerAgentTuiConfig,
+  WorkerAgentTuiConfigError,
+} from "./tui-config";
 import { createRemoteTuiDeliveryClient } from "./tui-remote";
 import {
   createTuiMessageSink,
   deliverRemoteTuiTurn,
   deliverTuiTurn,
   type TuiOutput,
-  WORKER_AGENT_TUI_CHANNEL,
 } from "./tui-sink";
-
-const TuiEnvironmentSchema = z
-  .object({
-    AI_API_KEY: z.string().optional(),
-    AI_BASE_URL: z.string().optional(),
-    AI_MODEL: z.string().optional(),
-    WORKER_AGENT_TUI_CHANNEL_ID: z.string().optional(),
-    WORKER_AGENT_TUI_DIR: z.string().optional(),
-    WORKER_AGENT_TUI_ENDPOINT: z.string().url().optional(),
-    WORKER_AGENT_TUI_TOKEN: z.string().optional(),
-  })
-  .passthrough();
-
-export type WorkerAgentTuiConfig =
-  | WorkerAgentLocalTuiConfig
-  | WorkerAgentRemoteTuiConfig;
-
-export interface WorkerAgentLocalTuiConfig {
-  readonly channel: ChannelAddress;
-  readonly directory: string;
-  readonly env: WorkerAgentModelEnv;
-  readonly mode: "local";
-}
-
-export interface WorkerAgentRemoteTuiConfig {
-  readonly channel: ChannelAddress;
-  readonly endpoint: string;
-  readonly mode: "remote";
-  readonly token?: string;
-}
 
 export interface StartWorkerAgentTuiOptions {
   readonly config?: WorkerAgentTuiConfig;
   readonly host?: AgentHost;
   readonly output?: TuiOutput;
-}
-
-export function resolveWorkerAgentTuiConfig(
-  environment: NodeJS.ProcessEnv = env,
-  args: readonly string[] = argv.slice(2)
-): WorkerAgentTuiConfig {
-  const parsed = TuiEnvironmentSchema.parse(environment);
-  const remoteEndpoint = readRemoteEndpoint(
-    parsed.WORKER_AGENT_TUI_ENDPOINT,
-    args
-  );
-  const channel = {
-    id:
-      parsed.WORKER_AGENT_TUI_CHANNEL_ID?.trim() || WORKER_AGENT_TUI_CHANNEL.id,
-    kind: "tui",
-  } satisfies ChannelAddress;
-
-  if (remoteEndpoint) {
-    return {
-      channel,
-      endpoint: remoteEndpoint,
-      mode: "remote",
-      ...(parsed.WORKER_AGENT_TUI_TOKEN?.trim()
-        ? { token: parsed.WORKER_AGENT_TUI_TOKEN.trim() }
-        : {}),
-    };
-  }
-
-  const apiKey = parsed.AI_API_KEY?.trim();
-  if (!apiKey) {
-    throw new WorkerAgentTuiConfigError(
-      "AI_API_KEY is required for local TUI mode. Set WORKER_AGENT_TUI_ENDPOINT or pass --remote for remote mode."
-    );
-  }
-
-  return {
-    channel,
-    directory:
-      parsed.WORKER_AGENT_TUI_DIR?.trim() ||
-      join(homedir(), ".pss-next", "worker-agent-tui"),
-    env: {
-      AI_API_KEY: apiKey,
-      ENVIRONMENT: "development",
-      ...(parsed.AI_BASE_URL?.trim()
-        ? { AI_BASE_URL: parsed.AI_BASE_URL.trim() }
-        : {}),
-      ...(parsed.AI_MODEL?.trim() ? { AI_MODEL: parsed.AI_MODEL.trim() } : {}),
-    },
-    mode: "local",
-  };
 }
 
 export async function startWorkerAgentTui(
@@ -136,6 +57,10 @@ export async function startWorkerAgentTui(
 
   output.writeLine("pss worker-agent TUI. Type /exit to quit.");
   input.setPrompt("> ");
+  if (inputClosed) {
+    await close.dispose();
+    return;
+  }
   input.prompt();
 
   for await (const line of input) {
@@ -175,22 +100,25 @@ async function configureTuiTurnDelivery(
           join(config.directory, "session-index.json")
         )
       );
-      const conversationKey = channelKey(config.channel);
+      const binding = localChannelBinding(config.channel);
       const agent = createConfiguredAgent(config.env, host, {
         sendMessage: {
           channel: () => config.channel,
           sink: createTuiMessageSink(output),
         },
         sessionTools: {
-          currentConversationKey: () => conversationKey,
+          currentConversationKey: () => binding.channelKey,
           reader: sessionIndex,
           transcriptReader: createThreadStoreSessionTranscriptReader({
-            resolveThreadKey: (conversationKey) => conversationKey,
+            resolveThreadKey: (conversationKey) =>
+              conversationKey === binding.channelKey
+                ? binding.threadKey
+                : conversationKey,
             store: threadStoreForHost(host),
           }),
         },
       });
-      const thread = agent.thread(channelKey(config.channel));
+      const thread = agent.thread(binding.thread);
       return {
         deliver: async (text) => {
           const assistantText: string[] = [];
@@ -204,7 +132,8 @@ async function configureTuiTurnDelivery(
           if (trimmed) {
             await sessionIndex.upsert({
               assistantText,
-              channel: config.channel,
+              channel: binding.channel,
+              threadKey: binding.threadKey,
               userText: trimmed,
             });
           }
@@ -230,21 +159,6 @@ async function configureTuiTurnDelivery(
   }
 }
 
-function readRemoteEndpoint(
-  environmentValue: string | undefined,
-  args: readonly string[]
-): string | undefined {
-  const flagValue = args.find((arg) => arg.startsWith("--remote="));
-  if (flagValue) {
-    return flagValue.slice("--remote=".length).trim();
-  }
-
-  const remoteFlagIndex = args.indexOf("--remote");
-  const remoteArgValue =
-    remoteFlagIndex >= 0 ? args[remoteFlagIndex + 1]?.trim() : undefined;
-  return remoteArgValue || environmentValue?.trim();
-}
-
 function isMainModule(moduleUrl: string, argvPath = argv[1]): boolean {
   return (
     argvPath !== undefined &&
@@ -260,11 +174,4 @@ function assertNever(value: never): never {
   throw new WorkerAgentTuiConfigError(
     `Unexpected TUI config variant: ${String(value)}`
   );
-}
-
-export class WorkerAgentTuiConfigError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "WorkerAgentTuiConfigError";
-  }
 }
