+++ packages/runtime/src/index.ts
@@ -8,6 +8,7 @@ export {
   type ThreadKey,
   type ThreadMetadata,
 } from "./agent/core/agent";
+export { threadStoreKey } from "./agent/core/thread-entry";
 export type { AgentHost } from "./execution/host/types";
 export type { AgentToolChoice } from "./llm/llm";
 export { delegateUserInput } from "./thread/input/delegate-input";
